Add multi user support to server
The server now spawns a thread for each user that connects.
It will display a message to all users that someone has joined and will
output messages recieved to all clients.

To compile the server use:
gcc server.c -o server -std=c99 -lpthread

The client must be adjusted so that it is always listening.